Added timeouts to claws-mail subprocess call.
So that if `claws-mail --status` freaks out, as it has every so often,
we'll kill it before it starts eating up an entire cpu core. We can then
try again in the next iteration.

This new feature requires python3.3, which isn't symlinked by debian's
'python3' yet. So manually specifying 3.3 in the shebang. We'll probably
want to switch that later once 3.3 is no longer on most systems. Also
put in a version check at startup.